Finishing up the H2 in-memory settings: * now has two columns UUID (primary key) and Date (creation dateAndTime) * upon uploading a research object the application now automatically renames the zip to a randomly created UUID and unzips it, also saves these details to the in-mem DB * refactored the class distribution in packages to better show their resposibilities
