+++ .gitignore
@@ -7,7 +7,7 @@ release.properties
 dependency-reduced-pom.xml
 buildNumber.properties
 .mvn/timing.properties
-.idea
+.idea/
 upload-dir/
 ro-show.iml
