+++ src/main/java/org/researchobject/roshow/controllers/RoController.java
@@ -1,6 +1,5 @@
-package org.researchobject.roshow;
+package org.researchobject.roshow.controllers;
 
-import java.io.File;
 import java.io.IOException;
 import java.util.stream.Collectors;
