+++ src/main/java/org/researchobject/roshow/model/UUIDdb.java
@@ -1,31 +1,28 @@
 package org.researchobject.roshow.model;
 
 
+import org.hibernate.annotations.Type;
+
 import javax.persistence.*;
 import java.io.Serializable;
-import java.time.LocalDate;
 import java.util.UUID;
 
 @Entity
 @Table(name = "uuid_table")
 public class UUIDdb implements Serializable {
 
-    private static final long serialVersionUID = -3009157732242241606L;
     @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private long id;
-
-    @Column(name = "uuid")
+    @Type(type="uuid-char")
     private UUID uuid;
 
     @Column(name = "date")
-    private LocalDate date;
+    private String date;
 
     public UUIDdb() {
 
     }
 
-    public UUIDdb(UUID uuid, LocalDate date) {
+    public UUIDdb(UUID uuid, String date) {
         this.uuid = uuid;
         this.date = date;
     }
@@ -39,11 +36,11 @@ public class UUIDdb implements Serializable {
     }
 
 
-    public LocalDate getDate() {
+    public String getDate() {
         return date;
     }
 
-    public void setDate(LocalDate date) {
+    public void setDate(String date) {
         this.date = date;
     }
