+++ src/main/java/org/researchobject/roshow/repository/UUIDrepository.java
@@ -1,4 +1,4 @@
-package org.researchobject.roshow.storage;
+package org.researchobject.roshow.repository;
 
 import java.util.List;
 import java.util.UUID;
@@ -6,6 +6,6 @@ import java.util.UUID;
 import org.researchobject.roshow.model.UUIDdb;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UUIDrepository extends CrudRepository<UUIDdb, Long>{
+public interface UUIDrepository extends CrudRepository<UUIDdb, UUID>{
     List<UUIDdb> findByUuid(UUID uuid);
 }
