+++ src/main/resources/application.properties
@@ -5,7 +5,7 @@ spring.thymeleaf.cache=false
 
 spring.h2.console.enabled=true
 spring.h2.console.path=/h2_console
-spring.datasource.url=jdbc:h2:file:~/h2/roshowDB
+spring.datasource.url=jdbc:h2:mem:roshowDB
 spring.datasource.username=sa
 spring.datasource.password=
 spring.datasource.driverClassName=org.h2.Driver
